Move OpenSSLTool to its own file
