+++ lib/heroku/command/certs.rb
@@ -1,4 +1,5 @@
 require "heroku/command/base"
+require "heroku/open_ssl_tool"
 require "excon"
 
 # manage ssl endpoints for an app
@@ -160,7 +161,7 @@ class Heroku::Command::Certs < Heroku::Command::Base
     domain = args[0] || error("certs:generate must specify a domain")
     subject = args[1]
     
-    keyfile, csrfile = OpenSSLTool.generate_csr(domain, subject)
+    keyfile, csrfile = Heroku::OpenSSLTool.generate_csr(domain, subject)
     
     display "Submit the CSR in #{csrfile} to your preferred certificate authority."
     display "When you've received your certificate, run:"
@@ -171,10 +172,10 @@ class Heroku::Command::Certs < Heroku::Command::Base
       display "$ heroku certs:add CERTFILE #{keyfile}"
     end
     
-  rescue OpenSSLTool::NotInstalledError => ex
+  rescue Heroku::OpenSSLTool::NotInstalledError => ex
     error("The OpenSSL command-line tools must be installed to use certs:generate.\n" + ex.installation_hint)
     
-  rescue OpenSSLTool::GenericError => ex
+  rescue Heroku::OpenSSLTool::GenericError => ex
     error(ex.message)
   end
   
@@ -260,48 +261,5 @@ class Heroku::Command::Certs < Heroku::Command::Base
     endpoints.select { |endpoint| endpoint['ssl_cert'] && endpoint['ssl_cert']['cert_domains'] } \
               .map   { |endpoint| endpoint['ssl_cert']['cert_domains'] } \
               .reduce(:+)
-  end
-  
-  module OpenSSLTool
-    def self.generate_csr(domain, subject = nil, key_size = 2048)
-      ensure_openssl_installed!
-      
-      keyfile = "#{domain}.key"
-      csrfile = "#{domain}.csr"
-    
-      subj_args = if subject
-        ['-subj', subject]
-      else
-        []
-      end
-    
-      system("openssl", "req", "-new", "-newkey", "rsa:#{key_size}", "-nodes", "-keyout", keyfile, "-out", csrfile, *subj_args) or raise GenericError, "Key and CSR generation failed: #{$?}"
-      
-      return [keyfile, csrfile]
-    end
-    
-    class GenericError < StandardError; end
-    
-    class NotInstalledError < GenericError
-      include Heroku::Helpers
-      
-      def installation_hint
-        if running_on_a_mac?
-          "With Homebrew <http://brew.sh> installed, run the following command:\n$ brew install openssl"
-        elsif running_on_windows?
-          "Download and install OpenSSL from <http://slproweb.com/products/Win32OpenSSL.html>."
-        else
-          # Probably some kind of Linux or other Unix. Who knows what package manager they're using?
-          "Make sure your package manager's 'openssl' package is installed."
-        end
-      end
-    end
-    
-  private
-    def self.ensure_openssl_installed!
-      return if @checked
-      system("openssl", "version") or raise NotInstalledError
-      @checked = true
-    end
   end
 end
